requirements.txt
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.5.0
streamlit>=1.24.0
